TSCN-Domain-Controller/requirement.txt
PyYAML
kombu
msgpack-python
redis
jenkinsapi
flask
flask-restful
Flask-Login
MySQL-python
setproctitle
gunicorn
gevent
Flask-SocketIO
docker-py
arrow